Separate test resource name helpers
